+++ src/common/text.c
@@ -643,7 +643,7 @@ get_stamp_str (char *fmt, time_t tim, char **ret)
 		return 0;
 	}
 
-	return len_utf8;
+	return (int) len_utf8;
 }
 
 static void
